// senado-agenda/web/AgendaWeb/ClientApp/src/app/models/espacio.model.ts
import { FuseUtils } from '@fuse/utils';

export class EspacioModel
{
  id: number;
  nombre: string;
  id_ubicacion: number
  capacidad: number;
  activo: number;
  nombre_ubicacion: string;
  calle_ubicacion: string;
  nombre_estado: string;
  nombre_municipio: string;
  nombre_localidad: string;
  en_uso: number;

  //constructor(espacioModel)
  //{
  //  {
  //    this.id = espacioModel.id || null;
  //    this.nombre = espacioModel.nombre || null;
  //    this.id_ubicacion = espacioModel.ubicacion || null;
  //    this.capacidad = espacioModel.capacidad || null;
  //    this.activo = espacioModel.activo || null;
  //    this.nombre_ubicacion = espacioModel.nombre_ubicacion || null;
  //  }
  //}
}